requirements.txt
Flask>=3.0.0
pytest>=8.0.0
firebase-admin>=6.0.0,<7
python-dotenv>=1.0.0
